Added configuration files to model_spec_helper
I'm most likely doing something wrong, but I kept getting this warning:

```
uninitialized constant <Model Name>::Rails (NameError)
```

when requiring model specs from ``model_spec_helper``. The tests would still run, but requiring this got rid of the warning. But then again it's loading Rails, which is the anti-thesis of your blog. With rails 4.1 and the ``Guardfile`` I referenced in #1 it's okay, but I don't think it's ideal for your case.